Move backup time display to the FileBrowser component
