+++ client/my-sites/backup/backup-contents-page/file-browser/index.tsx
@@ -1,6 +1,15 @@
-import { useState } from '@wordpress/element';
+import {
+	__experimentalText as Text,
+	__experimentalHStack as HStack,
+	ExternalLink,
+} from '@wordpress/components';
+import { createInterpolateElement, useState } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
 import { FunctionComponent } from 'react';
+import useGetDisplayDate from 'calypso/components/jetpack/daily-backup-status/use-get-display-date';
+import { useFirstMatchingBackupAttempt } from 'calypso/my-sites/backup/hooks';
 import { useSelector } from 'calypso/state';
+import { getSiteSlug } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import FileBrowserHeader from './file-browser-header';
 import FileBrowserNode from './file-browser-node';
@@ -13,6 +22,7 @@ export interface FileBrowserConfig {
 	alwaysInclude?: string[];
 	showHeaderButtons?: boolean;
 	showFileCard?: boolean;
+	showBackupTime?: boolean;
 	siteId?: number;
 }
 
@@ -32,6 +42,21 @@ const FileBrowser: FunctionComponent< FileBrowserProps > = ( {
 	const selectedSiteId = useSelector( getSelectedSiteId ) as number;
 	const effectiveSiteId = siteId ?? selectedSiteId;
 
+	const effectiveSiteSlug = useSelector( ( state ) => getSiteSlug( state, effectiveSiteId ) ) || '';
+	const getDisplayDate = useGetDisplayDate( effectiveSiteId );
+
+	const { backupAttempt: lastKnownBackupAttempt } = useFirstMatchingBackupAttempt(
+		effectiveSiteId,
+		{
+			sortOrder: 'desc',
+			successOnly: true,
+		}
+	);
+
+	const displayBackupDate = lastKnownBackupAttempt
+		? getDisplayDate( lastKnownBackupAttempt.activityTs, false )
+		: null;
+
 	const handleClick = ( path: string ) => {
 		setActiveNodePath( path );
 	};
@@ -49,6 +74,24 @@ const FileBrowser: FunctionComponent< FileBrowserProps > = ( {
 				showHeaderButtons={ fileBrowserConfig?.showHeaderButtons ?? true }
 				siteId={ effectiveSiteId }
 			/>
+			{ fileBrowserConfig?.showBackupTime && (
+				<HStack alignment="left" spacing={ 1 }>
+					<Text
+						color="var(--studio-gray-40)"
+						style={ { marginInlineStart: '14px', marginTop: '10px' } }
+					>
+						{ displayBackupDate
+							? createInterpolateElement( __( 'Listing files from the latest backup: <date />.' ), {
+									date: <span>{ displayBackupDate }</span>,
+							  } )
+							: __( 'There are no backups.' ) }{ ' ' }
+						<ExternalLink
+							href={ `/backup/${ effectiveSiteSlug }` }
+							children={ __( 'Create fresh backup now' ) }
+						/>
+					</Text>
+				</HStack>
+			) }
 			<FileBrowserNode
 				rewindId={ rewindId }
 				item={ rootItem }
