+++ client/sites/staging-site/components/staging-site-sync-modal/index.tsx
@@ -22,7 +22,6 @@ import {
 import { __, isRTL } from '@wordpress/i18n';
 import { error, chevronRight, chevronLeft } from '@wordpress/icons';
 import QueryRewindState from 'calypso/components/data/query-rewind-state';
-import useGetDisplayDate from 'calypso/components/jetpack/daily-backup-status/use-get-display-date';
 import InlineSupportLink from 'calypso/dashboard/components/inline-support-link';
 import { SectionHeader } from 'calypso/dashboard/components/section-header';
 import SiteEnvironmentBadge, {
@@ -57,6 +56,7 @@ const fileBrowserConfig: FileBrowserConfig = {
 	alwaysInclude: [ 'wp-config.php' ],
 	showHeaderButtons: false,
 	showFileCard: false,
+	showBackupTime: true,
 };
 
 const DirectionArrow = () => {
@@ -204,15 +204,12 @@ export default function SyncModal( {
 	const stagingSiteTitle = useSelector( ( state ) => getSiteTitle( state, stagingSiteId ) ) || '';
 
 	const targetSiteSlug = targetEnvironment === 'production' ? productionSiteSlug : stagingSiteSlug;
-	const sourceSiteSlug = sourceEnvironment === 'staging' ? stagingSiteSlug : productionSiteSlug;
 
 	const sourceSiteTitle = sourceEnvironment === 'staging' ? stagingSiteTitle : productionSiteTitle;
 	const targetSiteTitle =
 		targetEnvironment === 'production' ? productionSiteTitle : stagingSiteTitle;
 
 	const querySiteId = sourceEnvironment === 'staging' ? stagingSiteId : productionSiteId;
-
-	const getDisplayDate = useGetDisplayDate( querySiteId );
 
 	const browserCheckList = useSelector( ( state ) =>
 		getBackupBrowserCheckList( state, querySiteId )
@@ -372,10 +369,6 @@ export default function SyncModal( {
 		( showDomainConfirmation && domainConfirmation !== productionSiteSlug ) ||
 		( browserCheckList.totalItems === 0 && browserCheckList.includeList.length === 0 );
 
-	const displayBackupDate = lastKnownBackupAttempt
-		? getDisplayDate( lastKnownBackupAttempt.activityTs, false )
-		: null;
-
 	return (
 		<Modal
 			title={ syncConfig[ environment ].title }
@@ -455,22 +448,6 @@ export default function SyncModal( {
 							siteId={ querySiteId }
 							fileBrowserConfig={ fileBrowserConfig }
 						/>
-						<HStack alignment="left" spacing={ 1 }>
-							<Text color="var(--studio-gray-40)" style={ { marginInlineStart: '14px' } }>
-								{ displayBackupDate
-									? createInterpolateElement(
-											__( 'Listing files from the latest backup: <date />.' ),
-											{
-												date: <span>{ displayBackupDate }</span>,
-											}
-									  )
-									: __( 'There are no backups.' ) }{ ' ' }
-								<ExternalLink
-									href={ `/backup/${ sourceSiteSlug }` }
-									children={ __( 'Create fresh backup now' ) }
-								/>
-							</Text>
-						</HStack>
 					</div>
 					<HStack
 						alignment="left"
